Add GetAllCurrencies

+++ Emtia.Takip.Net6.Rest.Api/Model/HaremModel/USDKG.cs
@@ -7,8 +7,8 @@ namespace Emtia.Takip.Net6.Rest.Api.Helper
         public string satis { get; set; }
         public string tarih { get; set; }
         public Dir dir { get; set; }
-        public int dusuk { get; set; }
-        public int yuksek { get; set; }
-        public int kapanis { get; set; }
+        public double dusuk { get; set; }
+        public double yuksek { get; set; }
+        public double kapanis { get; set; }
     }
 }
